everything works within docker env, get route gets mock data from db, docker now runs spring server with docker compose up

+++ spring/src/main/java/spring/backend/models/JobRepository.java
@@ -2,6 +2,6 @@ package spring.backend.models;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobRepository extends JpaRepository<Job, Long> {
+public interface JobRepository extends JpaRepository<Job, Integer> {
     // Custom query methods can be added here
 }
